feat: separate routes for main application (#97)
