+++ ui/env.js
@@ -1,0 +1,2 @@
+// envs will be dynamically updated when the multiwoven server runs
+window.VITE_API_HOST = '__VITE_API_HOST__';
